fix: cross-platform disk usage and config permission checks
- Extract syscall.Statfs into disk_unix.go (linux/darwin/freebsd)
  with explicit uint64 casts for FreeBSD int64 fields
- Add disk_windows.go using PowerShell Get-PSDrive fallback
- Add disk_other.go stub for unsupported platforms (openbsd etc)
- Skip config permission check on Windows (0600 not enforceable)

+++ internal/commands/disk_unix.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build linux || darwin || freebsd
 
 package commands
 
@@ -15,8 +15,8 @@ func getDiskUsage(path string) (*diskUsage, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return nil, err
 	}
-	available := stat.Bavail * uint64(stat.Bsize)
-	total := stat.Blocks * uint64(stat.Bsize)
+	available := uint64(stat.Bavail) * uint64(stat.Bsize)
+	total := uint64(stat.Blocks) * uint64(stat.Bsize)
 	used := total - available
 	return &diskUsage{
 		Available:   available,
